Refactored project structure: separated core into multiple apps and cleaned migrations
