+++ buildlink/Users/views.py
@@ -1,10 +1,14 @@
-from django.shortcuts import render
 from rest_framework import generics, permissions, status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework_simplejwt.tokens import RefreshToken, TokenError
+
 from .models import User
-from .serializers import RegisterSerializer, UserSerializer
+from .serializers import (
+    RegisterSerializer,
+    UserSerializer,
+    WorkerProfileSerializer,
+)
 
 # Register API
 class RegisterView(generics.CreateAPIView):
@@ -13,7 +17,7 @@ class RegisterView(generics.CreateAPIView):
     permission_classes = [permissions.AllowAny]
 
 
-# User Profile API (current logged-in user)
+# Current-user profile (generic, any role)
 class UserProfileView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
@@ -22,13 +26,12 @@ class UserProfileView(APIView):
         return Response(serializer.data)
 
 
-# Logout API (invalidate token)
+# Logout API (blacklist refresh token)
 class LogoutView(APIView):
     permission_classes = [permissions.IsAuthenticated]
 
     def post(self, request):
         refresh_token = request.data.get("refresh")
-
         if not refresh_token:
             return Response({"detail": "Refresh token is required."}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -40,4 +43,10 @@ class LogoutView(APIView):
             return Response({"detail": "Invalid or expired token."}, status=status.HTTP_400_BAD_REQUEST)
 
 
-# Create your views here.
+# Worker profile (read/update own data + trades)
+class WorkerProfileView(generics.RetrieveUpdateAPIView):
+    serializer_class = WorkerProfileSerializer
+    permission_classes = [permissions.IsAuthenticated]
+
+    def get_object(self):
+        return self.request.user
