+++ buildlink/applications/models.py
@@ -1,0 +1,22 @@
+from django.db import models
+from Users.models import User
+from projects.models import Job
+from django.utils.translation import gettext_lazy as _
+
+class Application(models.Model):
+    class Status(models.TextChoices):
+        PENDING = 'pending', _('Pending')
+        ACCEPTED = 'accepted', _('Accepted')
+        REJECTED = 'rejected', _('Rejected')
+
+    job = models.ForeignKey(Job, on_delete=models.CASCADE, related_name='applications')
+    applicant = models.ForeignKey(User, on_delete=models.CASCADE, related_name='applications')
+    status = models.CharField(max_length=20, choices=Status.choices, default=Status.PENDING)
+    created_at = models.DateTimeField(auto_now_add=True)
+
+    class Meta:
+        unique_together = ('job', 'applicant')
+
+    def __str__(self):
+        return f"{self.applicant.full_name} applied to {self.job.title}"
+# Create your models here.
