+++ buildlink/buildlink/settings.py
@@ -39,7 +39,15 @@ INSTALLED_APPS = [
     'django.contrib.staticfiles',
     'rest_framework',
     'rest_framework_simplejwt',
-    'core',
+    'rest_framework_simplejwt.token_blacklist',
+   
+   
+    'Users',
+    'notifications',
+    'projects',
+    'trades',
+    'applications',
+    'ratings',
     
 ]
 
@@ -129,7 +137,7 @@ STATIC_URL = 'static/'
 
 DEFAULT_AUTO_FIELD = 'django.db.models.BigAutoField'
 
-AUTH_USER_MODEL = 'core.User'
+AUTH_USER_MODEL = 'Users.User'
 
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': (
